requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
colorama==0.3.9
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
Pillow==5.3.0
pluggy==0.7.1
py==1.6.0
pytest==3.8.2
six==1.11.0
Werkzeug==0.14.1
